cleanup, refactor and begin reworking certain sections
- removed:
  - MACCHINA_CONF handling
- refactored:
  - main
- reworked:
  - config handling
