+++ doc/README.md
@@ -3,7 +3,12 @@
 This directory contains documentation pertaining to the usage and configuration of _macchina_.
 
 |                       Document | Description                                                                                                                  |
-| -----------------------------: | :--------------------------------------------------------------------------------------------------------------------------- |
+|                           ---: | :---                                                                                                                         |
 | [macchina.toml](macchina.toml) | This is _macchina's_ configuration file, you can use this as a template and build on top of it.                              |
-| [macchina.adoc](macchina.adoc) | This is the document format we use to write our manpages, [Asciidoctor](https://asciidoctor.org/) is very easy to work with. |
-|       [macchina.1](macchina.1) | This is the 1st section of _macchina's_ manpages and the result of compiling [macchina.adoc](macchina.adoc).                 |
+
+---
+
+|                      Directory | Description                                                                                                                             |
+|                           ---: | :---                                                                                                                                    |
+|                   [adoc](adoc) | This directory contains the documents in which we write our manpages, which are then compiled and placed into the [man](man) directory  |
+|                     [man](man) | This directory contains macchina's manpages, i.e. the documents compiled from the [adoc](adoc) directory.                               |
