+++ doc/adoc/macchina.1.adoc
