+++ doc/adoc/macchina.2.adoc
@@ -1,0 +1,340 @@
+= macchina(2)
+Aziz Ben Ali, Marvin Haschker and Uttarayan Mondal
+v6.0.0
+:doctype: manpage
+:manmanual: MACCHINA
+:mansource: MACCHINA
+:man-linkstyle: pass:[blue R < >]
+
+== Name
+
+macchina - A system information fetcher, with an (unhealthy) emphasis on performance.
+
+== Synopsis
+
+*$XDG_CONFIG_HOME/macchina/themes*, *~/.config/macchina/themes*, *$MACCHINA_DIR/themes*
+
+== Description
+
+Themes are your interface to customizing all aspects of macchina. +
+This manpage details all the options a theme can accept.
+
+== General Options
+
+*spacing*::
+   Defines the amount of spacing to leave between +
+   the separator and the content besides it.
+
+   Takes an integer value, e.g. spacing = 1
+
+*padding*::
+   Defines the amount of padding to leave between +
+   the content and its surroundings.
+
+   Takes an integer value, e.g. padding = 0
+
+*hide_ascii*::
+   Disables the rendering of ASCII, whether it be +
+   built-in or custom.
+
+   Takes a boolean value, e.g. hide_ascii = false
+
+*prefer_small_ascii*::
+   For built-in ASCII, always use smaller variants.
+
+   Takes a boolean value, e.g. prefer_small_ascii = true
+
+*separator*::
+   Defines the glyph to use for the separator.
+
+   Takes a string value, e.g. separator = "-->"
+
+*key_color*::
+   Defines the color of the keys.
+
+   Takes hexadecimal/indexed/predefined color names,
+   where casing is insensitive e.g.
+
+   key_color = "#00FF00" or
+   key_color = "046" or
+   key_color = "Green"
+
+*separator_color*::
+   Defines the color of the separator.
+
+   Takes hexadecimal/indexed/predefined color names,
+   where casing is insensitive e.g.
+
+   separator_color = "#00FF00" or
+   separator_color = "046" or
+   separator_color = "Green"
+
+== Palette Section
+
+This section is optional, omitting it from the configuration +
+hides the palette completely.
+   
+*type*:: 
+
+Defines the type of colors to be displayed in the palette. + 
+The colors you configure your terminal to use in its colorschemes + 
+will be displayed by this component.
+
+   Takes a string value, e.g.
+
+   type = "Dark" or
+   type = "Light" or
+   type = "Full"
+
+*glyph*:: 
+
+Defines the glyph to use for the palette. You should +
+append a space to leave some room between the glyphs.
+
+   Takes a string value, e.g. glyph = "() "
+
+*visible*:: 
+
+Defines whether to show or hide the palette.
+
+   Takes a boolean value, e.g. visible = true
+
+== Bar Section:
+
+Bars are a good way to visualize your system usage without + 
+seeing the statistics, they range from 0% to 100%.
+
+This section is optional, omitting it from the configuration +
+hides the bars completely.
+
+*glyph*:: 
+
+Defines the glyph to use for all bars.
+
+   Takes a string value, e.g. glyph = "o"
+
+*symbol_open*::
+
+Defines the character to use for opening delimiters. Be sure +
+to surround the value with single quotes and not double quotes.
+
+   Takes a char value, e.g. symbol_open = '('
+
+*symbol_close*::
+
+Defines the character to use for closing delimiters. Be sure +
+to surround the value with single quotes and not double quotes.
+
+   Takes a char value, e.g. symbol_close = ')'
+
+*visible*::
+
+Defines whether to show or hide the bars.
+
+   Takes a boolean value, e.g. visible = true
+
+*hide_delimiters*::
+
+Defines whether to show or hide the bars delimiters, i.e. +
+the characters that surround the bars themselves.
+
+   Takes a boolean value, e.g. hide_delimiters = false
+
+== Box Section
+
+The box component renders a box that surrounds your system information.
+
+*title*::
+Defines the title of the box.
+
+   Takes a string value, e.g. title = "Hydrogen"
+
+*border*::
+Defines the type of border to use for the box.
+
+   Takes a string value, where casing is insensitive e.g.
+
+   border = "plain" or
+   border = "thick" or
+   border = "rounded" or
+   border = "double"
+
+*visible*::
+Defines whether to show or hide the box.
+
+   Takes a boolean value, e.g. visible = true
+
+== BOX.INNER_MARGIN Section
+
+*x*:: 
+Defines the horizontal margin to leave between + 
+the content and the box.
+
+   Takes an integer value, e.g. x = 2
+
+*y*::
+Defines the vertical margin to leave between + 
+the content and the box.
+
+   Takes an integer value, e.g. y = 1
+
+== CUSTOM_ASCII Section
+
+This section can be used to specify your own ASCII. +
+ANSI escape sequences are supported.
+
+*color*::
+Defines the color of the ASCII.
+
+   Takes hexadecimal/indexed/predefined color names, 
+   where casing is insensitive e.g.
+
+   separator_color = "#00FF00" or
+   separator_color = "046" or
+   separator_color = "Green"
+
+*path*::
+Defines the path to a file on your filesystem which contains
+the ASCII art you want to display.
+
+   Takes a string value, where casing is sensitive e.g.
+
+   path = "~/ascii/arch_linux"
+
+== Randomize Section
+
+This section is used to randomize color selection.
+
+*key_color*::
+Defines whether to randomize the color of the keys.
+
+   Takes a boolean value, e.g. key_color = true
+
+*separator_color*::
+Defines whether to randomize the color of the separator.
+
+   Takes a boolean value, e.g. separator_color = true
+
+*pool*::
+Defines the pool of colors from which to pick a random color.
+
+   Takes a string value, where casing is insensitive e.g.
+
+   pool = "hexadecimal" or
+   pool = "indexed" or
+   pool = "base"
+
+   If "hexadecimal" is specified, you'll get a random color ranging
+   from #000000 to #FFFFFF
+
+   If "indexed" is specified, you'll get a random color ranging
+   from 0 to 255
+
+   If "base" is specified, you'll get random color from the following
+   set of colors: "Black", "White", "Red", "Green", "Blue", "Yellow", 
+   "Magenta" and "Cyan".
+
+== Keys Section
+
+This section provides the options needed to +
+customize the keys of each readout.
+
+*host*::
+Defines the text of the Host readout.
+
+   Takes a string value, e.g. host = "Host"
+
+*kernel*::
+Defines the text of the Kernel readout.
+
+   Takes a string value, e.g. kernel = "Kernel"
+
+*os*::
+Defines the text of the OperatingSystem readout.
+
+   Takes a string value, e.g. os = "OS"
+
+*machine*::
+Defines the text of the Machine readout.
+
+   Takes a string value, e.g. machine= "Machine"
+
+*de*::
+Defines the text of the _DesktopEnvironment_ readout.
+
+   Takes a string value, e.g. de = "DE"
+
+*wm*::
+Defines the text of the _WindowManager_ readout.
+
+   Takes a string value, e.g. wm = "WM"
+
+*distro*::
+Defines the text of the Distribution readout.
+
+   Takes a string value, e.g. distro = "Distro"
+
+*terminal*::
+Defines the text of the Terminal readout.
+
+   Takes a string value, e.g. terminal = "Term"
+
+*shell*::
+Defines the text of the Shell readout.
+
+   Takes a string value, e.g. shell = "Shell"
+
+*packages*::
+Defines the text of the Packages readout.
+
+   Takes a string value, e.g. host = "Host"
+
+*uptime*::
+Defines the text of the Uptime readout.
+
+   Takes a string value, e.g. host = "Host"
+
+
+*local_ip*::
+Defines the text of the LocalIP readout.
+
+   Takes a string value, e.g. local_ip = "Local IP"
+
+*memory*::
+Defines the text of the Memory readout.
+
+   Takes a string value, e.g. memory = "Memory"
+
+*battery*::
+Defines the text of the Battery readout.
+
+   Takes a string value, e.g. battery = "Battery"
+
+*backlight*::
+Defines the text of the Backlight readout.
+
+   Takes a string value, e.g. backlight = "Brightness"
+
+*resolution*::
+Defines the text of the Resolution readout.
+
+   Takes a string value, e.g. resolution = "Resolution"
+
+*cpu*::
+Defines the text of the Processor readout.
+
+   Takes a string value, e.g. cpu = "CPU"
+
+*cpu_load*::
+Defines the text of the ProcessorLoad readout.
+
+   Takes a string value, e.g. cpu_load = "CPU %"
+
+== See Also
+
+macchina(1)
+
+== Copyright
+
+Copyright (c) 2021 Aziz Ben Ali +
+Distributed under the terms of the MIT license.
