+++ src/cli.rs
@@ -1,5 +1,6 @@
+use crate::config::Config;
 use crate::data;
-use clap::App;
+use crate::error;
 use serde::{Deserialize, Serialize};
 use std::default::Default;
 use structopt::StructOpt;
@@ -65,15 +66,6 @@ pub struct Opt {
     pub theme: Option<String>,
 
     #[structopt(
-        long = "export-config",
-        short = "e",
-        help = "Prints a template configuration file to stdout",
-        conflicts_with = "doctor"
-    )]
-    #[serde(skip_serializing, skip_deserializing)]
-    pub export_config: bool,
-
-    #[structopt(
         long = "list-themes",
         short = "l",
         help = "Lists all available themes: built-in and custom"
@@ -110,7 +102,6 @@ impl Default for Opt {
         Opt {
             version: false,
             doctor: false,
-            export_config: false,
             current_shell: false,
             long_shell: false,
             long_uptime: false,
@@ -126,7 +117,76 @@ impl Default for Opt {
     }
 }
 
-#[allow(dead_code)]
-pub fn build_cli() -> App<'static, 'static> {
-    Opt::clap()
+impl Opt {
+    pub fn parse_args(&mut self, args: Opt) {
+        if args.version {
+            self.version = true;
+        }
+
+        if args.doctor {
+            self.doctor = true;
+        }
+
+        if args.current_shell {
+            self.current_shell = true;
+        }
+
+        if args.long_shell {
+            self.long_shell = true;
+        }
+
+        if args.long_uptime {
+            self.long_uptime = true;
+        }
+
+        if args.list_themes {
+            self.list_themes = true;
+        }
+
+        if args.long_kernel {
+            self.long_shell = true;
+        }
+
+        if args.physical_cores {
+            self.physical_cores = true;
+        }
+
+        if args.ascii_artists {
+            self.ascii_artists = true;
+        }
+
+        if args.config.is_some() {
+            self.config = args.config;
+        }
+
+        if args.theme.is_some() {
+            self.theme = args.theme;
+        }
+
+        if args.show.is_some() {
+            self.show = args.show;
+        }
+
+        if args.interface.is_some() {
+            self.interface = args.interface;
+        }
+    }
+
+    pub fn get_options(arg_opt: Opt) -> Opt {
+        let config_opt = match arg_opt.config {
+            Some(_) => Config::read_config(&arg_opt.config.clone().unwrap()),
+            None => Config::get_config(),
+        };
+
+        match config_opt {
+            Ok(mut config) => {
+                config.parse_args(arg_opt);
+                config
+            }
+            Err(e) => {
+                error::print_errors(e);
+                arg_opt
+            }
+        }
+    }
 }
