+++ src/extra.rs
@@ -1,17 +1,10 @@
 use std::path::{Path, PathBuf};
 
-pub fn config_data_paths() -> [Option<PathBuf>; 4] {
-    [
-        dirs::config_dir(),
-        libmacchina::dirs::macos_config_dir(),
-        libmacchina::dirs::localbase_dir(),
-        libmacchina::dirs::usr_share_dir(),
-    ]
-}
-
+// Thanks to Andrey Tyukin
 // https://stackoverflow.com/questions/54267608/expand-tilde-in-rust-path-idiomatically
-pub fn expand_home<P: AsRef<Path>>(path_user_input: P) -> Option<PathBuf> {
-    let p = path_user_input.as_ref();
+pub fn expand_home<P: AsRef<Path>>(initial_path: P) -> Option<PathBuf> {
+    let p = initial_path.as_ref();
+
     if !p.starts_with("~") {
         return Some(p.to_path_buf());
     }
@@ -22,8 +15,6 @@ pub fn expand_home<P: AsRef<Path>>(path_user_input: P) -> Option<PathBuf> {
 
     dirs::home_dir().map(|mut h| {
         if h == Path::new("/") {
-            // Corner case: `h` root directory;
-            // don't prepend extra `/`, just drop the tilde.
             p.strip_prefix("~").unwrap().to_path_buf()
         } else {
             h.push(p.strip_prefix("~/").unwrap());
